Fix the backend logo sizing.
The former value of `100%` will show logos (especially) SVGs to large (this only happens in Safari for me).

The new value of `inherit` is much more what we want: a logo always fitted inside its boundary. In order to keep the ratio we need to set the `width` value as well.